adjust tuner settings

+++ src/uci.cpp
@@ -112,7 +112,7 @@ void UCI::Bench() {
     std::uint64_t end_time = get_time();
     long double total_time = 1.0 * (end_time - start_time) / 1000.0;
 
-    std::cout << total_nodes << " nodes " << (total_nodes / total_time)
+    std::cout << total_nodes << " nodes " << (long)(total_nodes / total_time)
               << " nps" << std::endl;
 }
